Fix memory replacement problem
When the variable is reassigned, it should be replaced completely
instead of changing pointer value due to different types of memory
objects (object and function) because during runtime variable
should be able to change its class from object to function and vise
versa. When we initialize object only once and then redefine its
value by pointer, we are not able to change its super class.